+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to the "vsfilemanager" extension will be documented in this 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.1] - 2025-10-15
+
+### Improved
+- Moved field edit notification from top of page to bottom right corner for better visibility
+
 ## [1.3.0] - 2025-10-15
 
 ### Added